requirements.txt
adium-theme-ubuntu==0.3.4
alabaster==0.7.7
Babel==1.3
backports.shutil-get-terminal-size==1.0.0
Beaker==1.9.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.6
django-debug-toolbar==1.8
djangorestframework==3.7.3
docutils==0.6
ecdsa==0.13
enum34==1.1.6
et-xmlfile==1.0.1
first==2.0.1
flake8==3.4.1
Flask==0.12.2
Flask-SocketIO==2.9.3
FormEncode==2.0.0a1
funcsigs==1.0.2
greenlet==0.4.9
gunicorn==19.7.1
gyp==0.1
idna==2.6
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mercurial==3.7.3
MySQL-python==1.2.5
mysqlclient==1.3.12
nose==1.3.7
olefile==0.44
openpyxl==2.4.9
paramiko==1.16.0
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
pathlib==1.0.1
pbr==3.1.1
pew==0.1.26
Pillow==4.3.0
pip-tools==1.9.0
pipenv==8.0.6
psycopg2==2.7.3.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.1
Pylons==1.0.2
python-apt==1.1.0b1
python-engineio==2.0.1
python-socketio==1.8.4
pythonz-bd==1.11.4
pytz==2017.2
-e git+https://github.com/diepmv/QuickWiki.git@7336fd8024916e5ea843caf2a3a99d577356b464#egg=QuickWiki
repoze.lru==0.7
requests==2.18.4
roman==2.0.0
Routes==2.4.1
shutilwhich==1.1.0
simplejson==3.13.2
six==1.11.0
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.1.14
sqlparse==0.2.4
stevedore==1.27.1
Tempita==0.5.3.dev0
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
waitress==1.1.0
WebError==0.13.1
WebHelpers==1.3
WebOb==1.7.4
WebTest==2.0.29
Werkzeug==0.12.2
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
